Add files via upload
The following code, written with my colleague morebaconstrips, implements the algebraic idea of the Symmetric Group.
Running the code you are given a choice: 
    1) Make the script calculate all the permutations of a group, inserting in input the cardinality of the group
    2) Insert a particular permutations, of which you want to calculate something

If your choice was the second one, you can choose now what operation to do on the permutation:
    1) calculate the inverse permutation
    2) calculate the product in disjoint cycles
    3) calculate the period of the permutation
    4) calculate the parity of the permutation
    5) calculate them all

This exercise was for us usefull preparing to the Algebra exam in our CS course. 
Have fun manipulating permutations!